tests/dialog-enabled.html

<!DOCTYPE html>
<html>
<head>
<script src="../dist/index.js"></script>
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
</head>
<body>
<dialog></dialog>
<script>
  var dialogs = document.querySelectorAll('dialog');

  for (var i = 0; i < dialogs.length; i++) {
      new AyDialog(dialogs[i]);
  }
</script>
<script>
test(function() {
  dialog = document.querySelector('dialog')
  assert_true(dialog instanceof HTMLDialogElement);
}, "The DIALOG element should be recognized");
</script>
</body>
</html>
